requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-multipart==0.0.6
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
pillow>=9.0.0
einops>=0.6.0
python-jose>=3.3.0
